♻️  refactor: update file structure
